Add template part context to navigation block
