+++ packages/block-library/src/navigation-overlay-close/index.js
@@ -2,18 +2,16 @@
  * WordPress dependencies
  */
 import { addFilter } from '@wordpress/hooks';
-import { select } from '@wordpress/data';
-import { store as coreStore } from '@wordpress/core-data';
 
 /**
  * Internal dependencies
  */
 import initBlock from '../utils/init-block';
+import { isWithinNavigationOverlay } from '../utils/is-within-overlay';
 import edit from './edit';
 import metadata from './block.json';
 import save from './save';
 import icon from './icon';
-import { NAVIGATION_OVERLAY_TEMPLATE_PART_AREA } from '../navigation/constants';
 
 const { name } = metadata;
 
@@ -24,40 +22,6 @@ export const settings = {
 	edit,
 	save,
 };
-
-function isWithinOverlay() {
-	// @wordpress/block-library should not depend on @wordpress/editor.
-	// Blocks can be loaded into a *non-post* block editor, so to avoid
-	// declaring @wordpress/editor as a dependency, we must access its
-	// store by string.
-	// eslint-disable-next-line @wordpress/data-no-store-string-literals
-	const editorStore = select( 'core/editor' );
-
-	// Return false if the editor store is not available.
-	if ( ! editorStore ) {
-		return false;
-	}
-
-	const { getCurrentPostType, getCurrentPostId } = editorStore;
-	const { getEditedEntityRecord } = select( coreStore );
-
-	const postType = getCurrentPostType();
-	const postId = getCurrentPostId();
-
-	if ( postType === 'wp_template_part' && postId ) {
-		const templatePartEntity = getEditedEntityRecord(
-			'postType',
-			'wp_template_part',
-			postId
-		);
-
-		return (
-			templatePartEntity?.area === NAVIGATION_OVERLAY_TEMPLATE_PART_AREA
-		);
-	}
-
-	return false;
-}
 
 export const init = () => {
 	addFilter(
@@ -72,7 +36,7 @@ export const init = () => {
 				return canInsert;
 			}
 
-			return isWithinOverlay();
+			return isWithinNavigationOverlay();
 		}
 	);
